+++ Changelog.md
@@ -30,6 +30,7 @@ v2.36.0 (not published yet)
 - NVEncC: Add support for Dolby Vision metadata handling
 - NVEncC: Fix wrong "--vpp-resize" parameter options visibility
 - QSVEncC: Add support for Dolby Vision metadata handling
+- QSVEncC: Add "Show advanced QP settings"
 - VCEEncC: Add support for Dolby Vision metadata handling
 - VCEEncC: Add Dolby Vision parameters to encoder options
 - VCEEncC: Fix "--cqp" parameter value for HEVC codec ([#1222](/../../issues/1222))